+++ packages/components/organisms/f-footer/test-utils/component-objects/f-footer.selectors.js
@@ -1,5 +1,13 @@
-export const FEEDBACK_BUTTON = '[data-test-id="feedback-button"]';
-export const COUNTRY_SELECTOR_BUTTON = '[data-test-id="countrySelector-button"]';
-export const COUNTRY_LINK = 'data-test-id="countrySelector-country-';
-export const FOOTER_ICON = 'data-test-id="footerIcon ';
-export const FOOTER_LINK_LIST = '[data-test-id="footer-link-list"]';
+const FEEDBACK_BUTTON = '[data-test-id="feedback-button"]';
+const COUNTRY_SELECTOR_BUTTON = '[data-test-id="countrySelector-button"]';
+const COUNTRY_LINK = 'data-test-id="countrySelector-country-';
+const FOOTER_ICON = 'data-test-id="footerIcon ';
+const FOOTER_LINK_LIST = '[data-test-id="footer-link-list"]';
+
+export {
+    FEEDBACK_BUTTON,
+    COUNTRY_SELECTOR_BUTTON,
+    COUNTRY_LINK,
+    FOOTER_ICON,
+    FOOTER_LINK_LIST
+};