+++ packages/components/organisms/f-footer/test/visual/f-footer.visual.desktop.spec.js
@@ -1,57 +1,63 @@
-import forEach from 'mocha-each';
+import Footer from '../../test-utils/component-objects/f-footer.component';
 
-const Footer = require('../../test-utils/component-objects/f-footer.component');
-
-let footer;
+let locales;
 
 describe('f-footer - Desktop Visual Tests', () => {
-    beforeEach(() => {
-        // Arrange
-        footer = new Footer();
+    locales = [
+        'en-GB',
+        'en-AU',
+        'en-IE',
+        'en-NZ',
+        'es-ES',
+        'it-IT'
+    ];
+
+    locales.forEach(locale => {
+        it(`should display the base footer for tenant: "${locale}"`, async () => {
+            // Act
+            await Footer.load({ locale });
+
+            // Assert
+            await browser.percyScreenshot(`f-footer - Base - ${locale}`, 'desktop');
+        });
+
+        it(`should display the footer with country selector for tenant: "${locale}"`, async () => {
+            // Act
+            await Footer.load({ locale, showCountrySelector: true });
+
+            // Assert
+            await browser.percyScreenshot(`f-footer - Country Selector - ${locale}`, 'desktop');
+        });
+
+        it(`should display slim footer when no content links are provided for tenant: "${locale}"`, async () => {
+            // Act
+            await Footer.load({ locale, showLinksContent: false });
+
+            // Assert
+            await browser.percyScreenshot(`f-footer - Slim Footer - ${locale}`, 'desktop');
+        });
     });
 
-    forEach(['en-GB', 'en-AU', 'en-IE', 'en-NZ', 'es-ES', 'it-IT'])
-        .it('should display the base footer for tenant: "%s"', async tenant => {
+    locales = [
+        'en-AU',
+        'en-NZ'
+    ];
+
+    locales.forEach(locale => {
+        it(`should display the footer with courier links for tenant: "${locale}"`, async () => {
             // Act
-            await footer.load({ locale: tenant });
+            await Footer.load({ locale, showCourierLinks: true });
 
             // Assert
-            await browser.percyScreenshot(`f-footer - Base - ${tenant}`, 'desktop');
+            await browser.percyScreenshot(`f-footer - Courier Links - ${locale}`, 'desktop');
         });
 
-    forEach(['en-GB', 'en-AU', 'en-IE', 'en-NZ', 'es-ES', 'it-IT'])
-        .it('should display the footer with country selector for tenant: "%s"', async tenant => {
+        it(`should display the footer with courier links and country selector for tenant: "${locale}"`, async () => {
             // Act
-            await footer.load({ locale: tenant, showCountrySelector: true });
+            await Footer.load({ locale, showCountrySelector: true, showCourierLinks: true });
 
             // Assert
-            await browser.percyScreenshot(`f-footer - Country Selector - ${tenant}`, 'desktop');
+            await browser.percyScreenshot(`f-footer - Courier Links and Country Selector - ${locale}`, 'desktop');
         });
-
-    forEach(['en-GB', 'en-AU', 'en-IE', 'en-NZ', 'es-ES', 'it-IT'])
-        .it('should display slim footer when no content links are provided for tenant: "%s"', async tenant => {
-            // Act
-            await footer.load({ locale: tenant, showLinksContent: false });
-
-            // Assert
-            await browser.percyScreenshot(`f-footer - Slim Footer - ${tenant}`, 'desktop');
-        });
-
-    forEach(['en-AU', 'en-NZ'])
-        .it('should display the footer with courier links for tenant: "%s"', async tenant => {
-            // Act
-            await footer.load({ locale: tenant, showCourierLinks: true });
-
-            // Assert
-            await browser.percyScreenshot(`f-footer - Courier Links - ${tenant}`, 'desktop');
-        });
-
-    forEach(['en-AU', 'en-NZ'])
-        .it('should display the footer with courier links and country selector for tenant: "%s"', async tenant => {
-            // Act
-            await footer.load({ locale: tenant, showCountrySelector: true, showCourierLinks: true });
-
-            // Assert
-            await browser.percyScreenshot(`f-footer - Courier Links and Country Selector - ${tenant}`, 'desktop');
-        });
+    });
 });